More progress

+++ src/main/java/at/hannibal2/skyhanni/config/storage/ProfileSpecificStorage.kt
@@ -577,7 +577,7 @@ class ProfileSpecificStorage {
             var attemptsSince: Int = 0
 
             @Expose
-            var xpSince: Int = 0
+            var xpSince: Long = 0
         }
 
         @Expose
